[Fuzzing] Fix oss-fuzz build and use a different file for benchmark testing
