+++ udp/client/mutexmap_test.go
@@ -70,12 +70,3 @@ func TestMutexMap(t *testing.T) {
 		t.Logf("Failed, outLists: %#v", outLists)
 	}
 }
-
-func BenchmarkM(b *testing.B) {
-	m := NewMutexMap()
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		// run uncontended lock/unlock - should be quite fast
-		m.Lock(i).Unlock()
-	}
-}
