+++ udp/coder/coder_test.go
@@ -195,54 +195,6 @@ func TestUnmarshalMessage(t *testing.T) {
 	})
 }
 
-func BenchmarkMarshalMessage(b *testing.B) {
-	options := make(message.Options, 0, 32)
-	bufOptions := make([]byte, 1024)
-	bufOptionsUsed := bufOptions
-
-	enc := 0
-	options, enc, _ = options.SetPath(bufOptionsUsed, "/a/b/c/d/e")
-	bufOptionsUsed = bufOptionsUsed[enc:]
-
-	options, _, _ = options.SetContentFormat(bufOptionsUsed, message.TextPlain)
-	msg := message.Message{
-		Code:    codes.GET,
-		Payload: []byte{0x1},
-		Token:   []byte{0x1, 0x2, 0x3},
-		Options: options,
-	}
-	buffer := make([]byte, 1024)
-
-	b.ResetTimer()
-	for i := uint32(0); i < uint32(b.N); i++ {
-		_, err := DefaultCoder.Encode(msg, buffer)
-		if err != nil {
-			b.Fatalf("cannot marshal")
-		}
-	}
-}
-
-func BenchmarkUnmarshalMessage(b *testing.B) {
-	buffer := []byte{
-		0x40, 0x1, 0x30, 0x39, 0x46, 0x77,
-		0x65, 0x65, 0x74, 0x61, 0x67, 0xa1, 0x3,
-		0xff, 'h', 'i',
-	}
-	options := make(message.Options, 0, 32)
-	msg := message.Message{
-		Options: options,
-	}
-
-	b.ResetTimer()
-	for i := uint32(0); i < uint32(b.N); i++ {
-		msg.Options = options
-		_, err := DefaultCoder.Decode(buffer, &msg)
-		if err != nil {
-			b.Fatalf("cannot unmarshal: %v", err)
-		}
-	}
-}
-
 func FuzzDecode(f *testing.F) {
 	f.Add([]byte{0x40, 0x1, 0x30, 0x39, 0x46, 0x77, 0x65, 0x65, 0x74, 0x61, 0x67, 0xa1, 0x3, 0xff, 'h', 'i'})
 	f.Add([]byte{88, 128, 107, 170, 134, 237, 158, 132, 150, 19, 19, 159, 72, 20, 210, 14, 23, 231, 160, 183, 145, 128, 177, 14, 82, 20, 210, 255})
